COMPOSE-351, iOS Simplify Swift code in examples
 - change indent from tabs to spaces
 - simplify and inline SwiftUI ContentView structure

https://youtrack.jetbrains.com/issue/COMPOSE-351/remove-redundant-struct-ContentView
