+++ examples/interop/ios-compose-in-swiftui/iosApp/iosApp/ComposeViewControllerInSwiftUI.swift
@@ -3,6 +3,15 @@ import SwiftUI
 import shared
 import MapKit
 
+struct ComposeLayer: View {
+    var body: some View {
+        ScreenTemplate(title: "Compose in SwiftUI") {
+            ComposeViewControllerInSwiftUI()
+                    .ignoresSafeArea(.keyboard) // Compose has its own keyboard handler
+        }
+    }
+}
+
 struct ComposeViewControllerInSwiftUI: UIViewControllerRepresentable {
     func makeUIViewController(context: Context) -> UIViewController {
         return Main_iosKt.ComposeEntryPoint()
