+++ examples/interop/ios-compose-in-swiftui/iosApp/iosApp/iOSApp.swift
@@ -1,0 +1,20 @@
+import SwiftUI
+
+@main
+struct iOSApp: App {
+    var body: some Scene {
+        WindowGroup {
+            TabView {
+                ComposeInSwiftUIScreen()
+                    .tabItem {
+                        Label("Compose in SwiftUI", systemImage: "star.fill")
+                    }
+
+                YetAnotherSwiftUIScreen()
+                    .tabItem {
+                        Label("SwiftUI", systemImage: "gear")
+                    }
+            }
+        }
+    }
+}
