--- examples/interop/ios-compose-in-swiftui/iosApp/iosApp/iosApp.swift
@@ -1,17 +1,0 @@
-import SwiftUI
-
-@main
-struct iOSApp: App {
-	var body: some Scene {
-		WindowGroup {
-			TabView {
-                ComposeInSwiftUIScreen()
-                    .tabItem { Label("Compose in SwiftUI", systemImage: "star.fill") }
-
-                YetAnotherSwiftUIScreen()
-                    .tabItem { Label("SwiftUI", systemImage: "gear") }
-                
-            }
-		}
-	}
-}
