+++ examples/interop/ios-swiftui-in-compose/README.md
@@ -1,7 +1,7 @@
 # SwiftUI in Compose
 
 This example shows how to use a SwiftUI view in a Compose application.
-Pay attention to file [main.ios.kt](shared%2Fsrc%2FiosMain%2Fkotlin%2Fmain.ios.kt) and [ComposeViewControllerToSwiftUI.swift](iosApp%2FiosApp%2FComposeViewControllerToSwiftUI.swift)
+Pay attention to file [main.ios.kt](shared%2Fsrc%2FiosMain%2Fkotlin%2Fmain.ios.kt) and [ComposeViewControllerRepresentable.swift](iosApp%2FiosApp%2FComposeViewControllerRepresentable.swift)
 
 ![screenshot.png](screenshot.png)
