--- examples/interop/ios-uikit-in-compose/iosApp/iosApp/iosApp.swift
@@ -1,13 +1,0 @@
-import SwiftUI
-import shared
-import UIKit
-
-@main
-struct iOSApp: App {
-	var body: some Scene {
-		WindowGroup {
-            ComposeViewControllerToSwiftUI()
-					.ignoresSafeArea(.all)
-		}
-	}
-}
